requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
Django==2.0.5
django-cleanup==2.1.0
django-ipware==2.1.0
kombu==4.1.0
netifaces==0.10.7
pyotp==2.2.6
pypng==0.0.18
PyQRCode==1.2.1
pytz==2018.4
redis==2.10.6
six==1.11.0
vine==1.1.4
